project/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
urllib3==2.0.2
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
urllib3==2.0.2
asgiref==3.6.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
urllib3==2.0.2
asgiref==3.6.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
urllib3==2.0.2
whitenoise==6.4.0
asgiref==3.6.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
urllib3==2.0.2
